# Expose sequence methods via CPython interface tables.